Fixed some minor warnings. Updated code for Visual Studio.

+++ librbr/include/core/rewards/sa_rewards_array.h
@@ -168,12 +168,12 @@ private:
 	/**
 	 * The minimum R-value.
 	 */
-	float Rmin;
+	double Rmin;
 
 	/**
 	 * The maximum R-value.
 	 */
-	float Rmax;
+	double Rmax;
 
 };
