+++ librbr/include/core/rewards/sas_rewards_array.h
@@ -152,12 +152,12 @@ private:
 	/**
 	 * The minimum R-value.
 	 */
-	float Rmin;
+	double Rmin;
 
 	/**
 	 * The maximum R-value.
 	 */
-	float Rmax;
+	double Rmax;
 
 };
